desktop-agent/requirements.txt
pyserial~=3.5
requests~=2.28.0
watchdog~=2.1.9
pydicom~=2.3.0
python-hl7~=0.4.5
pyinstaller~=5.7.0
